8th commit;complete the category 2 class linkage

+++ WebRoot/admin/productlist.jsp
@@ -59,7 +59,7 @@
     		<td><%=p.getMemberPrice() %></td>
     		<td><%=p.getPdate() %></td>
     		<td><%=Category.getCategoryById(p.getCategoryId()).getName() %></td>
-    		<td><a href="productdelete.jsp?id=<%=p.getId() %>" target="detail">É¾³ý</a>
+    		<td><a href="productdelete.jsp?id=<%=p.getId() %>" target="detail" onclick="return confirm('È·¶¨É¾³ý£¿')">É¾³ý</a>
     		|
     		<a href="productmodify.jsp?id=<%=p.getId() %>&categoryid=<%=p.getCategoryId() %>" target="detail">ÐÞ¸Ä</a>
     		|
